Revert entropy costs from uint64_t to float (restore v1.3.2 precision)
v1.6.0 changed all entropy costs to 23-bit fixed-point (LOG_2_PRECISION_BITS=23)
which causes a 10-19% lossless encode regression at method=4 vs v1.3.2. The
different precision produces different histogram merge decisions in the combine
phase, degrading output file size.

Changes across 12 files:
- kLog2Table, kSLog2Table: uint32_t/uint64_t → float (direct log2 values)
- VP8LFastLog2, VP8LFastSLog2: return float instead of fixed-point integers
- VP8LBitEntropy.entropy: float
- VP8LHistogram.bit_cost, costs[5]: float
- All entropy accumulation in histogram_enc.c, predictor_enc.c,
  backward_references_enc.c, backward_references_cost_enc.c,
  vp8l_enc.c: DivRound/SaturateAdd/<<LOG_2_PRECISION_BITS removed
- CombinedShannonEntropy: SSE2/AVX2/AVX-512 return float
  (AVX-512 keeps integer SIMD internally, converts at boundary)
- MAX_BIT_COST (1.e38f) replaces WEBP_UINT64_MAX as infinity sentinel

Keeps all v1.6.0 structural improvements: is_used[], costs[5] array,
trivial_symbol[5], swap-and-pop removal, AVX-512 dispatch.

Precision test (tests/test_float_entropy.c): 560/560 pass.
Max relative error: 5.07e-08 (kLog2Table), 5.22e-08 (kSLog2Table).

+++ src/dsp/lossless.h
@@ -193,9 +193,9 @@ extern VP8LPredictorAddSubFunc VP8LPredictorsSub_SSE[16];
 // Huffman-cost related functions.
 
 typedef uint32_t (*VP8LCostFunc)(const uint32_t* population, int length);
-typedef uint64_t (*VP8LCombinedShannonEntropyFunc)(const uint32_t X[256],
-                                                   const uint32_t Y[256]);
-typedef uint64_t (*VP8LShannonEntropyFunc)(const uint32_t* X, int length);
+typedef float (*VP8LCombinedShannonEntropyFunc)(const uint32_t X[256],
+                                                const uint32_t Y[256]);
+typedef float (*VP8LShannonEntropyFunc)(const uint32_t* X, int length);
 
 extern VP8LCostFunc VP8LExtraCost;
 extern VP8LCombinedShannonEntropyFunc VP8LCombinedShannonEntropy;
@@ -207,7 +207,7 @@ typedef struct {        // small struct to hold counters
 } VP8LStreaks;
 
 typedef struct {            // small struct to hold bit entropy results
-  uint64_t entropy;         // entropy
+  float entropy;            // entropy
   uint32_t sum;             // sum of the population
   int nonzeros;             // number of non-zero elements in the population
   uint32_t max_val;         // maximum value in the population
